+++ JollyLights.Tests/TestShopItemRepository.cs
@@ -14,7 +14,7 @@ namespace Jolly_Lights.Tests
             _tempDir = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString());
             Directory.CreateDirectory(_tempDir);
 
-            string testSchemaPath = Path.Combine(_tempDir, "schema.sql");     
+            string testSchemaPath = Path.Combine(_tempDir, "schema.sql");
 
             var originalSchemaPath = Path.Combine(AppContext.BaseDirectory, "..", "..", "..", "..", "Jolly-Lights-Cinema-Group", "Database", "schema.sql");
             File.Copy(originalSchemaPath, testSchemaPath);
@@ -84,9 +84,20 @@ namespace Jolly_Lights.Tests
         [TestCleanup]
         public void Cleanup()
         {
-            if (Directory.Exists(_tempDir))
+            if (_tempDir != null && Directory.Exists(_tempDir))
             {
-                Directory.Delete(_tempDir, recursive: true);
+                for (int i = 0; i < 5; i++)
+                {
+                    try
+                    {
+                        Directory.Delete(_tempDir, recursive: true);
+                        break;
+                    }
+                    catch (IOException)
+                    {
+                        Thread.Sleep(1); // If cinemaDB is still busy, then wait 1ms and try again
+                    }
+                }
             }
         }
     }
